tabs/withdrawal/explorer: move explorer to deposit/withdraw page
