+++ packages/sidesail/lib/pages/tabs/testchain/mainchain/deposit_withdraw_tab_route.dart
@@ -67,6 +67,10 @@ class DepositWithdrawTabPage extends StatelessWidget {
                             ),
                         ],
                       ),
+                      const DashboardGroup(
+                        title: 'Withdrawal Explorer',
+                        children: [WithdrawalExplorer()],
+                      ),
                     ],
                   ),
                 ],
